Authentication is tested and works properly
