+++ data/dbConfig.js
@@ -1,4 +1,5 @@
 const knex = require('knex');
-const knexConfig = require('../knexfile');
+const config = require('../knexfile');
+const environment = process.env.DB_ENV || 'development';
 
-module.exports = knex(knexConfig.development);
+module.exports = knex(config[environment]);
