--- data/middleware/generateToken.js
@@ -1,15 +1,0 @@
-const jwt = require('jsonwebtoken');
-const secrets = require('../secrets/secret');
-
-export default function generateToken(user) {
-    const payload = {
-        subject: user.id,
-        username: user.username,
-    }
-
-    const options = {
-        expiresIn: '1d'
-    }
-
-    return jwt.sign(payload, secrets.jwtSecret, options);
-}
