+++ knexfile.js
@@ -14,21 +14,6 @@ module.exports = {
       directory: './data/seeds',
     },
   },
-
-//   testing: {
-//     client: 'sqlite3',
-//     connection: {
-//       filename: './data/test.db3',
-//     },
-//     useNullAsDefault: true,
-//     migrations: {
-//       directory: './data/migrations',
-//     },
-//     seeds: {
-//       directory: './data/seeds',
-//     },
-//   },
-
   production: {
     client: 'pg',
     connection: dbConnnection,
